remembering boxing

+++ OCPLearning/src/pkgs/pkgExes/Exe001.java
@@ -18,10 +18,9 @@ public class Exe001 {
 		//System.out.println("[="+()+"]"+"");
 	}
 
-	short m1() {return 32767;}
-	//short m2() {return 32768;}//Type mismatch: cannot convert from int to short
-	short m3() {return new Short((short)5);}
-	Short m4() {return (short)5;}
+	long m1() {return 32767;}
+	long m3() {return new Long(5);}
+	Long m4() {return (long)5;}
 
 	public void im1(String[] args) {
 		System.out.println("Exe001.im1()");
